+++ src/main/java/com/booking/flight/utils/FareUtils.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class FareUtils {
 	
-	public Double calculateTotalFare(Double fare, Double numberOfPassengers) {
+	public Double calculateTotalFare(Double fare, Integer numberOfPassengers) {
 		return (fare*numberOfPassengers);
 	}
